+++ packages/scoutgame/src/builderNfts/registration/createBuilderNft.ts
