+++ packages/scoutgame/src/builderNfts/registration/createBuilderNftStarterPack.ts
