+++ packages/scoutgame/src/builderNfts/registration/registerDeveloperStarterNFT.ts
